Add 5-light support. Add patterns. Break API.
5-light support means that API consumers will
need to change how they use it, so I took the
opportunity to change the method signature more
than just changing to a pointer, with output
coming first in the signature in my preferred
style.

Patterns are a bit wonky, and they haven't been
nailed down in their design yet. Not sure how
preferable the design is. Dislike the weirdness
and some of the inflexibility - would be much
more nice with 16 bits rather than 8. Will need
to ask about that.