Relicense state machine; Increase modularization
Broke up dependent bits so they aren't all cluttering the same file. The
same process still needs to be applied to the remainder of the
statemachine file if it can be usefully detangled without issue.

I hereby license past and future contributions to GGreenwood's Shinewave
project under the MIT License.